Remove the 'const' qualifier from the C mapping of _maximum in bounded sequence
Having 'const' struct members in C++ is a bit difficult to handle in practice.
This deletes the default assignment operator and even though one was provided,
this prevents the genom_sequence_resize() from working correctly with sequences
of bounded sequences (because of the use of realloc(3)).

Since the C mappings can (and are) also be used from C++ code, the mapping must
be consistent for both laguages. The easiest solution is to remove the const
qualifier from the _maximum member.