Provide object context definition in [$object loc] template commands
For components, tasks, services, codels and parameters, the template
[$object loc] command returns a list of 4 elements instead of 3.
The 4th element corresponds to the 'component context' (a TCL component
object) where the object was originally defined and may be different
from the regular [$object component] in case of inheritance via
interfaces.

A new optional 'element' keyword is also added in [$object loc ?elem?]
for convenience. Element may be one of file, line, column or context
keyword and the corresponding information is returned instead of the
full list.