Skeleton template generates codels stubs only for codels directly defined
The skeleton template will skip codel stub generation for any codel defined in
external interfaces by default.

New skeleton options are added to change the default behaviour:
 -x generates stubs for all codels (previous behaviour)
 --extern=comp,... generates stubs for codels defined in those
 components/interfaces only.

This change should not affect existing components that do not use codels
defined in interfaces. In case a component is affected, the fix is simple and
consists in invoking the 'skeleton' in merge mode with the new -x option.