For ports provided via an interface, use the port type from the interface
Recreating a port type for interface provided ports has the inconvenient that
the original port in the interface and the one in the component appears to be
different. This creates an articifical dichotomy between codels that would be
defined by the interface and in the component. So it is more correct to reuse
the same original port type.

For backward compatibility with existing codels, an alias to the original type
is still created in the component namespace, so this change will be fully
transparent. In addition, simply merging new skeleton codel stub will update
the codels signature appropriately.

For ports 'used' via an interface, a new type must still be created in the
component namespace: as the direction is swapped, the type is different.