Fix -Wmaybe-uninitialized for gcc>=10
gcc>=10 now assumes that an enum can have any value, not just the defined
ones. This leads to many warnings after 'switch(enum)' constructs that have
no default case on purpose (to trigger -Wswitch warning).

To fix this, while preserving the -Wswitch behaviour, add a default case that
aborts and also force -Werror=switch-enum when using gcc (-Wswitch is silenced
when there is a default case). In the few cases where not all enum values must
be handled in a switch, cast the value to int to silence -Wswitch-enum.