Allow simple codels in attributes
There is no fundamental reason why an attribute could not have a simple
codel. Existing templates don't treat attributes and functions differently,
so it should be safe to allow simple codels.

This can be useful as a hook performing a related action when an attribute
is set. It's probably useless for attributes that have only out parameters,
but it does not hurt either.