Add a template helper function to parse templates' command line
Builtin skeleton template now can be given files or components names in order
to generate skeleton for a subset of components defined in all input files.
Since this functionality will also be useful in non-builtin templates, provide
helper functions for that.

The new template functions are (see documentation for details):
template arg files: the list of files arguments on the command line
template arg names: the list of arguments that are not files
template arg components: the components objects corresponding to names

Previous `template arg` function is unchanged, but can now be invoked as
`template arg next`.