Fix "return-value" generator function
Fix "language::return-value" that was not addressing the case of arguments of
type bounded string or array passed by value. E.g. an argument of type
"const type [n]" decays to "const type *" when returned from a function.

language::return-value was only used with arguments passed by reference in
existing templates, to this issue was overlooked for quite some time.

While here, are optional "var" argument, passed down to "language::argument"
when needed.